+++ src/web_server/request_handler.c
@@ -249,7 +249,7 @@ void request_handler_proxy(struct mg_connection *nc, struct mg_http_message *hm,
         //decode uri
         uri_decoded = sds_urldecode(uri_decoded, query, sdslen(query), false);
         if (is_allowed_proxy_uri(uri_decoded) == true) {
-            create_backend_connection(nc, backend_nc, uri_decoded, forward_backend_to_frontend);
+            create_backend_connection(nc, backend_nc, uri_decoded, forward_backend_to_frontend_stream);
         }
         else {
             webserver_send_error(nc, 403, "Host is not allowed");