+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This minor release adds support of new MPD 0.24 features. The javascript fronten
 - Feat: jsDoc compatible API documentation and linting with typescript compiler
 - Feat: add view action to default radio click actions
 - Feat: add "Add to" button to filesystem root #860
+- Feat: cache coverimages for webradio favorites
 - Upd: libmympdclient 1.0.15
 - Upd: refactor translation framework
 - Upd: en-US, de-DE translations